refactor(models): split ORM facade
Move Base, runtime engine setup, civic records, event records, and catalog records into focused modules while keeping pipeline.models as the public import surface.

Preserve one metadata registry and add tests for facade exports, table registration, and internal module boundaries.
